refactor(chat): 채팅 로직들을 훅으로 분리
